move non-plugin code in core plugins to own package
