+++ plugins/core/src/commonMain/kotlin/com/episode6/mxo2/plugins/core/CorePlugins.kt
@@ -1,6 +1,7 @@
 package com.episode6.mxo2.plugins.core
 
 import com.episode6.mxo2.MockspressoBuilder
+import com.episode6.mxo2.plugins.core.reflect.reflectionRealObjectMaker
 import com.episode6.mxo2.reflect.allConstructors
 import com.episode6.mxo2.reflect.asKClass
 import com.episode6.mxo2.reflect.parameterCount
